rename hex_utils
