+++ src/bech32_test.c
@@ -1,10 +1,8 @@
 #ifdef DEVEL
 
-#include <string.h>
-#include <os.h>
-#include "test_utils.h"
-#include "hex_utils.h"
 #include "bech32.h"
+#include "hexUtils.h"
+#include "test_utils.h"
 
 void testcase_bech32(const char* hrp, const char* inputBytesHex, const char* expectedStr)
 {
