+++ src/getSerial.h
@@ -1,7 +1,7 @@
 #ifndef H_CARDANO_APP_GET_SERIAL
 #define H_CARDANO_APP_GET_SERIAL
 
-#include <os.h>
+#include "handlers.h"
 
 handler_fn_t getSerial_handleAPDU;
