+++ src/hexUtils.c
@@ -1,9 +1,5 @@
-#include "errors.h"
-#include <stdint.h>
-#include <os.h>
+#include "common.h"
 #include "test_utils.h"
-#include <string.h>
-#include "utils.h"
 
 uint8_t hex_parseNibble(const char c)
 {
