+++ src/hexUtils.h
