+++ src/main.c
@@ -24,8 +24,8 @@
 #include <stdbool.h>
 #include <os_io_seproxyhal.h>
 #include <os.h>
-#include "ux.h"
 
+#include "ux.h"
 #include "getVersion.h"
 #include "handlers.h"
 #include "state.h"
@@ -33,7 +33,6 @@
 #include "menu.h"
 #include "assert.h"
 #include "io.h"
-#include "endian.h"
 
 // The whole app is designed for a specific api level.
 // In case there is an api change, first *verify* changes
