+++ src/runTests.c
@@ -1,14 +1,11 @@
 #ifdef DEVEL
 
-#include "common.h"
-
 #include "runTests.h"
 #include "cbor.h"
-#include "endian.h"
 #include "base58.h"
 #include "bech32.h"
 #include "test_utils.h"
-#include "hex_utils.h"
+#include "hexUtils.h"
 #include "hash.h"
 #include "keyDerivation.h"
 #include "addressUtilsByron.h"
